feat: update ResourceKey to include resource type index
BREAKING CHANGE:
The internal structure of `ResourceKey` has been modified to include a
resource type index.
 - Upper 16 bits: PAK file index in the AssetLoader's collection
 - Next 16 bits: Resource type index (index in ResourceTypeList)
 - Lower 32 bits: ResourceIndexT from the PAK format (resource index within the
 PAK)

 This allows efficient lookup and type-safe handling of loader and
 unloader functions for a resource using just is key.

+++ projects/Oxygen.Engine/src/Oxygen/Content/AssetLoader.h
@@ -379,7 +379,7 @@ private:
   OXGN_CNTT_API static auto HashResourceKey(const ResourceKey& key) -> uint64_t;
 
   //! Get PAK file index from pointer (for resource key creation)
-  auto GetPakIndex(const PakFile& pak) const -> uint32_t;
+  auto GetPakIndex(const PakFile& pak) const -> uint16_t;
 
   //! Helper: Create asset loader call with unified error handling
   template <LoadFunction F>
